\name{running_results}
\alias{running_results}
\docType{data}
\title{Results of all runs completed}
\description{Times and distances of all completed runs}
\usage{data("running_results")}
\format{
  A data frame with the following 3 variables.
  \itemize{
    \item training_date. the date the run was completed
    \item distance_in_meters. the distance completed in meters
    \item time_in_seconds. the time in seconds taken to complete the run
  }
}
\details{None}
\source{}
\references{}
\examples{
data(running_results)
## maybe str(running_results) ; plot(running_results) ...
}
\keyword{datasets}
